WIP added mapped_variables
this provides a full long format list of file x variable construct x variable (with definition) list. Intent is to use to quickly combine variables of multiple files and select variables of interest within files. TODO: delete geographic indicators (eg. GEOID) and add a join key. Ultimately, this will be converted into a json file and potentially added to metadata although not necessary (and can be used locally as well)